+++ gatsby-config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  siteMetadata: {
+    title: 'Ali Ukani',
+  },
+  plugins: [
+    'gatsby-plugin-emotion',
+    'gatsby-plugin-react-helmet',
+  ],
+}
