+++ package.json
@@ -3,9 +3,9 @@
   "version": "2.1.0",
   "description": "site served at ali.io",
   "scripts": {
-    "dev": "next",
-    "build": "next build",
-    "start": "next start"
+    "start": "serve ./public --single",
+    "dev": "gatsby develop",
+    "build": "gatsby build"
   },
   "repository": {
     "type": "git",
@@ -18,8 +18,13 @@
   },
   "homepage": "https://github.com/ali/ali.io#readme",
   "dependencies": {
-    "next": "2.4.7",
+    "emotion": "^6.0.3",
+    "gatsby": "^1.2.0",
+    "gatsby-link": "^1.0.9",
+    "gatsby-plugin-emotion": "^1.1.0",
+    "gatsby-plugin-react-helmet": "^1.0.1",
     "react": "^15.6.1",
-    "react-dom": "^15.6.1"
+    "react-dom": "^15.6.1",
+    "serve": "^6.0.2"
   }
 }
